fix(codex): bound Computer Use process repair
